moved two way auth api route to two-way-auth folder
