--- apps/next/app/two-way-auth-enter/page.tsx
@@ -1,15 +1,0 @@
-'use client'
-
-import TwoWayAuthEnter from './TwoWayAuthEnter'
-import Content from '@/components/Content'
-import Section from '@/components/Section'
-
-const TwoWayAuthEnterPage = () => (
-	<Content>
-		<Section>
-			<TwoWayAuthEnter />
-		</Section>
-	</Content>
-)
-
-export default TwoWayAuthEnterPage
