+++ apps/next/app/two-way-auth/enter/page.tsx
@@ -1,0 +1,13 @@
+import TwoWayAuthEnter from './TwoWayAuthEnter'
+import Content from '@/components/Content'
+import Section from '@/components/Section'
+
+export default function TwoWayAuthEnterPage() {
+	return (
+		<Content>
+			<Section>
+				<TwoWayAuthEnter />
+			</Section>
+		</Content>
+	)
+}
